+++ src/types/postDataType.ts
